+++ tests/test_solver.py
@@ -1,6 +1,6 @@
 import pytest
-from src.utagmsengine.solver import Solver
-from src.utagmsengine.dataclasses import Preference, Indifference, Criterion, Position
+from src.utagmsengine.solver import Solver, Inconsistency
+from src.utagmsengine.dataclasses import Preference, Indifference, Criterion, Position, Intensity
 
 
 @pytest.fixture()
@@ -43,7 +43,17 @@ def predefined_criterions_dummy():
 
 @pytest.fixture()
 def positions_dummy():
-    return [Position(alternative_id='A', worst_position=6, best_position=1)]
+    return [Position(alternative_id='A', worst_position=12, best_position=2)]
+
+
+@pytest.fixture()
+def intensities_dummy():
+    return [Intensity(alternative_id_1='H', alternative_id_2='G', alternative_id_3='B', alternative_id_4='D', criteria=['g1', 'g2'])]
+
+
+@pytest.fixture()
+def resolved_inconsistencies_dummy():
+    return [[[], [], [Position(alternative_id='A', worst_position=12, best_position=9, criteria=[])], []], [[Preference(superior='G', inferior='F', criteria=[])], [], [], []], [[], [Indifference(equal1='D', equal2='G', criteria=[])], [], []]]
 
 
 @pytest.fixture()
@@ -53,12 +63,12 @@ def hasse_diagram_dict_dummy():
 
 @pytest.fixture()
 def representative_value_function_dict_dummy():
-    return {'J': 0.0113, 'E': 0.1057, 'C': 0.1614, 'F': 0.2559, 'K': 0.3116, 'B': 0.3248, 'H': 0.3731, 'L': 0.374, 'A': 0.4617, 'D': 0.5233, 'G': 0.5233, 'I': 0.6325}
+    return {'J': 0.0338, 'E': 0.1663, 'C': 0.196, 'F': 0.3015, 'K': 0.3311, 'B': 0.3344, 'H': 0.3972, 'L': 0.3972, 'A': 0.5323, 'D': 0.5756, 'G': 0.5756, 'I': 0.6227}
 
 
 @pytest.fixture()
 def criterion_functions_dummy():
-    return {'g1': [(0.0, 0.00342706), (2.0, 0.00440622), (6.0, 0.00636454), (7.0, 0.0), (9.0, 0.00783327), (16.0, 0.0112603), (18.0, 0.0424458), (25.0, 0.151595), (26.0, 0.167187), (35.0, 0.307522), (39.0, 0.369893), (62.0, 0.369893)], 'g2': [(2.0, 0.0), (9.0, 0.0), (14.666666666666664, 0.0), (15.0, 5.79318e-05), (17.0, 0.000405522), (24.0, 0.00162209), (30.0, 0.00266486), (38.33333333333333, 0.00411316), (40.0, 0.00411316), (43.0, 0.00411316), (55.0, 0.00411316), (62.0, 0.00411316)], 'g3': [(0.0, 0.0), (12.0, 0.101619), (14.0, 0.118555), (17.0, 0.14396), (25.0, 0.211706), (33.33333333333333, 0.282275), (44.0, 0.290399), (66.66666666666666, 0.307663), (68.0, 0.320396), (73.0, 0.368146), (88.0, 0.511395), (100.0, 0.625994)]}
+    return {'g1': [(0.0, 0.0102816), (2.0, 0.0132191), (6.0, 0.0190943), (7.0, 0.0), (9.0, 0.0235007), (16.0, 0.0337822), (18.0, 0.0606056), (25.0, 0.154487), (26.0, 0.167899), (35.0, 0.288604), (39.0, 0.342251), (62.0, 0.342251)], 'g2': [(2.0, 0.0), (9.0, 0.0), (14.666666666666664, 0.0), (15.0, 0.000773555), (17.0, 0.00541489), (24.0, 0.0216596), (30.0, 0.0355836), (38.33333333333333, 0.0549224), (40.0, 0.0549224), (43.0, 0.0549224), (55.0, 0.0549224), (62.0, 0.0549224)], 'g3': [(0.0, 0.0), (12.0, 0.11142), (14.0, 0.12999), (17.0, 0.157845), (25.0, 0.232125), (33.33333333333333, 0.309501), (44.0, 0.309501), (66.66666666666666, 0.309501), (68.0, 0.321234), (73.0, 0.365232), (88.0, 0.497229), (100.0, 0.602826)]}
 
 
 @pytest.fixture()
@@ -72,6 +82,7 @@ def test_get_hasse_diagram_dict(
         indifferences_dummy,
         criterions_dummy,
         positions_dummy,
+        intensities_dummy,
         hasse_diagram_dict_dummy
 ):
     solver = Solver(show_logs=True)
@@ -81,7 +92,8 @@ def test_get_hasse_diagram_dict(
         preferences_dummy,
         indifferences_dummy,
         criterions_dummy,
-        #positions_dummy
+        positions_dummy,
+        intensities_dummy
     )
 
     assert hasse_diagram_list == hasse_diagram_dict_dummy
@@ -94,23 +106,32 @@ def test_get_representative_value_function_dict(
         predefined_criterions_dummy,
         criterions_dummy,
         positions_dummy,
+        intensities_dummy,
         representative_value_function_dict_dummy,
-        criterion_functions_dummy
+        criterion_functions_dummy,
+        resolved_inconsistencies_dummy
 ):
     solver = Solver(show_logs=True)
 
-    representative_value_function_dict, criterion_functions, sampler_metrics = solver.get_representative_value_function_dict(
-        performance_table_dict_dummy,
-        preferences_dummy,
-        indifferences_dummy,
-        predefined_criterions_dummy,
-        [],
-        'files/polyrun-1.1.0-jar-with-dependencies.jar',
-        '10'
-    )
+    try:
+        representative_value_function_dict, criterion_functions, sampler_metrics = solver.get_representative_value_function_dict(
+            performance_table_dict_dummy,
+            preferences_dummy,
+            indifferences_dummy,
+            predefined_criterions_dummy,
+            positions_dummy,
+            intensities_dummy,
+            'files/polyrun-1.1.0-jar-with-dependencies.jar',
+            '10'
+        )
 
-    assert representative_value_function_dict == representative_value_function_dict_dummy
-    assert criterion_functions == criterion_functions_dummy
+        assert representative_value_function_dict == representative_value_function_dict_dummy
+        assert criterion_functions == criterion_functions_dummy
+
+    except Inconsistency as e:
+        resolved_inconsistencies = e.data
+
+        assert resolved_inconsistencies == resolved_inconsistencies_dummy
 
 
 def test_predefined_get_hasse_diagram_dict(